refactor: moved things around into client, server, test folders
