+++ src/main.rs
@@ -1,8 +1,11 @@
 use std::env::set_var;
-mod seq;
+mod test;
 mod utils;
+mod server;
+mod client;
+use test::run_seq::run;
 
 fn main() {
   set_var("RUST_BACKTRACE", "1");
-  seq::run();
+  run();
 }
