--- src/seq/README.md
@@ -1,4 +1,0 @@
-### Sequential Implementation
-
-The sequential implementation will use one thread. It will constantly just wait for the TCP
-listener to receive new messages
