--- src/seq/mod.rs
@@ -1,7 +1,0 @@
-mod clientsocket;
-mod testclient;
-pub mod seq;
-mod server;
-
-pub use super::utils::*;
-pub use seq::*;
