+++ src/server/mod.rs
@@ -1,0 +1,3 @@
+pub mod server;
+
+pub use server::*;
